rrhh ok

+++ rrhh-model/rrhh-model-jpa/src/main/java/org/sistcoop/rrhh/models/jpa/JpaSucursalProvider.java
@@ -7,18 +7,18 @@ import javax.ejb.Local;
 import javax.ejb.Stateless;
 import javax.ejb.TransactionAttribute;
 import javax.ejb.TransactionAttributeType;
-import javax.inject.Inject;
 import javax.inject.Named;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 
+import org.sistcoop.rrhh.models.ModelDuplicateException;
 import org.sistcoop.rrhh.models.SucursalModel;
 import org.sistcoop.rrhh.models.SucursalProvider;
+import org.sistcoop.rrhh.models.jpa.entities.AgenciaEntity;
 import org.sistcoop.rrhh.models.jpa.entities.SucursalEntity;
 import org.sistcoop.rrhh.models.search.SearchCriteriaModel;
 import org.sistcoop.rrhh.models.search.SearchResultsModel;
-import org.sistcoop.rrhh.models.search.filters.SucursalFilterProvider;
 
 @Named
 @Stateless
@@ -28,9 +28,6 @@ public class JpaSucursalProvider extends AbstractHibernateStorage implements Suc
 
     @PersistenceContext
     private EntityManager em;
-
-    @Inject
-    private SucursalFilterProvider filterProvider;
 
     @Override
     protected EntityManager getEntityManager() {
@@ -44,6 +41,12 @@ public class JpaSucursalProvider extends AbstractHibernateStorage implements Suc
 
     @Override
     public SucursalModel create(String denominacion) {
+        if (findByDenominacion(denominacion) != null) {
+            throw new ModelDuplicateException(
+                    "SucursalEntity denominacion debe ser unico, se encontro otra entidad con denominacion="
+                            + denominacion);
+        }
+
         SucursalEntity sucursalEntity = new SucursalEntity();
         sucursalEntity.setDenominacion(denominacion);
         em.persist(sucursalEntity);
@@ -52,6 +55,14 @@ public class JpaSucursalProvider extends AbstractHibernateStorage implements Suc
 
     @Override
     public boolean remove(SucursalModel sucursalModel) {
+        TypedQuery<AgenciaEntity> query = em.createNamedQuery("AgenciaEntity.findByIdSucursal",
+                AgenciaEntity.class);
+        query.setParameter("idSucursal", sucursalModel.getId());
+        query.setMaxResults(1);
+        if (!query.getResultList().isEmpty()) {
+            return false;
+        }
+
         SucursalEntity sucursalEntity = em.find(SucursalEntity.class, sucursalModel.getId());
         em.remove(sucursalEntity);
         return true;
@@ -64,19 +75,31 @@ public class JpaSucursalProvider extends AbstractHibernateStorage implements Suc
     }
 
     @Override
-    public SearchResultsModel<SucursalModel> search() {
+    public SucursalModel findByDenominacion(String denominacion) {
+        TypedQuery<SucursalEntity> query = em.createNamedQuery("SucursalEntity.findByDenominacion",
+                SucursalEntity.class);
+        query.setParameter("denominacion", denominacion);
+        List<SucursalEntity> results = query.getResultList();
+        if (results.isEmpty()) {
+            return null;
+        } else if (results.size() > 1) {
+            throw new IllegalStateException("Mas de una SucursalEntity con denominacion=" + denominacion
+                    + ", results=" + results);
+        } else {
+            return new SucursalAdapter(em, results.get(0));
+        }
+    }
+
+    @Override
+    public List<SucursalModel> getAll() {
         TypedQuery<SucursalEntity> query = em
                 .createNamedQuery("SucursalEntity.findAll", SucursalEntity.class);
 
         List<SucursalEntity> entities = query.getResultList();
-        List<SucursalModel> models = new ArrayList<SucursalModel>();
+        List<SucursalModel> result = new ArrayList<SucursalModel>();
         for (SucursalEntity bovedaEntity : entities) {
-            models.add(new SucursalAdapter(em, bovedaEntity));
+            result.add(new SucursalAdapter(em, bovedaEntity));
         }
-
-        SearchResultsModel<SucursalModel> result = new SearchResultsModel<>();
-        result.setModels(models);
-        result.setTotalSize(models.size());
         return result;
     }
 
@@ -97,7 +120,7 @@ public class JpaSucursalProvider extends AbstractHibernateStorage implements Suc
     @Override
     public SearchResultsModel<SucursalModel> search(SearchCriteriaModel criteria, String filterText) {
         SearchResultsModel<SucursalEntity> entityResult = findFullText(criteria, SucursalEntity.class,
-                filterText, filterProvider.getDenominacionFilter());
+                filterText, "denominacion");
 
         SearchResultsModel<SucursalModel> modelResult = new SearchResultsModel<>();
         List<SucursalModel> list = new ArrayList<>();
